more structural subgraph changes
