+++ opensea-subgraph-dev/schema.graphql
@@ -1,27 +1,15 @@
-# Subgraph Schema: NFT Marketplace
-# Version: 1.0.0
-
 enum NftStandard {
   ERC721
   ERC1155
   UNKNOWN
 }
 
-enum SaleStrategy {
-  " Strategy that executes an order at a fixed price that can be taken either by a bid or an ask. "
-  STANDARD_SALE
+enum SaleKind {
+  " Sell-side order, i.e. a buyer directly purchases an sell order. "
+  DIRECT_PURCHASE
 
-  " Strategy that executes an order at a fixed price that can be matched by any tokenId for the collection. "
-  ANY_ITEM_FROM_COLLECTION
-
-  " Strategy that executes an order at a fixed price that can be matched by any tokenId in a set of tokenIds. "
-  ANY_ITEM_FROM_SET
-
-  " Strategy to launch a Dutch Auction for a token where the price decreases linearly until a specified timestamp and end price defined by the seller. "
-  DUTCH_AUCTION
-
-  " Strategy to set up an order that can only be executed by a specific address. "
-  PRIVATE_SALE
+  " Buy-side order, i.e. a buyer place a bid at the auction, where the sale prices decreases overt time "
+  AUCTION
 }
 
 type Collection @entity {
@@ -37,7 +25,7 @@ type Collection @entity {
   " Total supply of the collection, mirrored from the smart contract. "
   totalSupply: BigInt
 
-  " NFT Standard the collection uses. "
+  " NFT Standard the collection uses. ERC721 or ERC1155. "
   nftStandard: NftStandard!
 
   " Royalty fee rate in percentage. E.g. 2.5% should be 2.5 "
@@ -111,7 +99,7 @@ type Trade @entity {
   priceETH: BigDecimal!
 
   " Stretegy that the trade is executed. "
-  strategy: SaleStrategy!
+  saleKind: SaleKind!
 
   " Buyer account address "
   buyer: String!
